LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-same-document-then-cross-document-expected.txt



FAIL navigation.activation same-document then cross-document assert_equals: expected 3 but got 1
